libc/picolibc: Split hooks into separate files
This splits the picolibc helper functions into separate files so that
code not explicitly referenced can be excluded from the result once
we stop using --whole-archive with this library.
