+++ boards/qemu/x86/qemu_x86_tiny.ld
@@ -144,10 +144,10 @@ MEMORY
 #endif /* CONFIG_NEWLIB_LIBC */
 
 #ifdef CONFIG_PICOLIBC
-/* For Picolibc libc-hook.c. */
+/* For Picolibc, all files under lib/libc/picolibc */
 #define LIB_C_IN_SECT(lsect)						\
-	*liblib__libc__picolibc.a:libc-hooks.c.obj(.##lsect)		\
-	*liblib__libc__picolibc.a:libc-hooks.c.obj(.##lsect##.*)
+	*liblib__libc__picolibc.a:(.##lsect)				\
+	*liblib__libc__picolibc.a:(.##lsect##.*)
 
 #endif /* CONFIG_PICOLIBC */
